# ENH : instanciator performances