After reading this
https://blogs.msdn.microsoft.com/seteplia/2017/02/01/dissecting-the-new-constraint-in-c-a-perfect-example-of-a-leaky-abstraction/

I gave a try at IL for RDD instanciator.
"Unfortunately", aspnetcore reflection and expression in general are much more performant than .net framework.

So the expression form is better/similar than IL, but way clearer.
Anyway, I proposed a different instanciator 10X more performant. The expression compilation cost is cashed only once since this component is registered as a singleton.


Method |      Mean |     Error |    StdDev | Scaled |  Gen 0 | Allocated |
----------------------- |----------:|----------:|----------:|-------:|-------:|----------:|
Lambda | 11.279 ns | 0.1079 ns | 0.0901 ns |   0.17 | 0.0102 |      32 B |
DefaultInstanciator | 64.811 ns | 0.7400 ns | 0.6560 ns |   1.00 | 0.0101 |      32 B |
ExpressionInstanciator |  6.791 ns | 0.1733 ns | 0.1621 ns |   0.10 | 0.0102 |      32 B |
ILInstanciator |  8.325 ns | 0.2062 ns | 0.2375 ns |   0.13 | 0.0102 |      32 B |